FIX: Update bcftools and tabix binaries
The python3.9 lambda environment is different to the python3.6
environment, and the binaries no longer work.

Recompile tabix and bcftools to run in the new environment, which
doesn't include the required libraries. Note this is a work in progress,
as tabix at least still doesn't recognise the libraries, and samtools
hasn't been updated yet.