FIX: Send max one last_batch=True from queryVCF
When queryVCF runs out of time publishing messages for queryGTF, it
batches up the remaining coordinates and sends them to the intermediate
queryVCFsubmit. However, each of the batches inherits the last_batch
value, which causes concat to be called for each batch.

Only propagate last_batch to the first batch of the split. We do this
instead of the last because all the SNSs will be published nearly
simultaneously and it is simpler to check if you are at the start of an
iteration than the end.