REFACTOR: Combine queryVCF and queryVCFExtended
The code and logic for queryVCF and queryVCFExtended are almost
identical, except for a slight difference in the recurse trigger and the
way in which each function was called.

Split queryVCF into its component parts; the API interface and the VCF
querying part. Roll the vcf querying into the (now renamed)
queryVCFExtended function, and create a new function, initQuery, to
handle the API logic.