REFACTOR: Always batch up queryVCF recursion
When queryVCF runs out of time it sends the rest of it's data bacck to
itself using SNS. It either sends the whole set of remaining data in one
message, or, if there is too much data, it sends the rest as several
smaller batches. This makes the function more complex than it needs to
be, as batching from the start improves parallelism anyway.

Always batch the recursion calls, as this reduces function complexity
and improves total pipeline speed.