REFACTOR: Call concatStarter from initQuery
The concept of last_batch is passed through the entire pipeline so
ideally the last plugin to run then calls concatStarter. This has been a
source of several bugs and leaves extra work for each function and to
add new plugins.

Now that concatStarter has been developed to cost a minimal amount of
money while waiting for the temp files to be destroyed, we can move it
to the start to remove all the last_batch logic from all the pipeline
functions. In addition, standardise the management of temporary files to
ensure that future additions are as simple as possible and make the
existence of temporary files watertight while the pipeline is running.