FEAT: Return presigned result url on /submit
Although jobs can be started using the /submit enpoint, there is no way
for users to know where or when the result is available, or even to
access it if they did.

Generate a presigned get url for the location of the result in advance
and return it. Change the format of the return to json with {"Response":
"Process started", "ResultUrl": presigned_url}. The presigned url is
valid for 24 hours.