REFACTOR: Reduce max run time from 8 to 2 hours
concatStarter waits for 8 hours for the component files to be completed
before starting the concatenation process. If the components take this
long either the workload is much larger than we'd like or there was a
crash early on somewhere.

Reduce the maximum waiting time for concatStarter to 2 hours, to reduce
the worst-case scenario and more quickly identify errors.