DEV: Update pluginConsequence module to build from a docker image
Provided runtimes were updated, deprecating the old version of this
module. To avoid this in future, a docker container has been created
that builds the equivalent as a docker image and uses that to deploy
the lambda function.

* Added docker_images.tf and docker_prep.py - defines the ecr and
computes hash on files to determine any changes to the image

* Updated lambda function in main.tf - No longer using fork of
terraform-aws-lambda, though this will likely be changed once
support is added

* Updated sns.tf to reflect attributes from non-forked lambda source

* Added variables.tf with region as it is needed to define ecr

* Updated versions.tf - made aws version explicit and added docker

* Added Dockerfile - configures required environment for VEP

* Added bootstrap.sh - handles lambda function initiation and passing
of sns notifications to VEP

* Updated VEP.pm handler to support new sns notification passing

* Updated VEP.pm handler to reference location awscli is installed to
in docker container

* Updated VEP.pm to only write to .tsv when the results array contains
elements